\name{writing_functions-package}
\alias{writing_functions-package}
\alias{writing_functions}
\docType{package}
\title{
\packageTitle{writing_functions}
}
\description{
\packageDescription{writing_functions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{writing_functions}
\packageIndices{writing_functions}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{writing_functions}

Maintainer: \packageMaintainer{writing_functions}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
